+++ EssentialFeed.xcodeproj/xcuserdata/gopal.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -36,8 +36,8 @@
             filePath = "EssentialFeedTests/Feed API/RemoteFeedImageDataLoaderTests.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "159"
-            endingLineNumber = "159"
+            startingLineNumber = "100"
+            endingLineNumber = "100"
             landmarkName = "test_loadImageDataFromURL_doesNotDeliverResultAfterCancellingTask()"
             landmarkType = "7">
          </BreakpointContent>
